Revert "Scrum 244 backend adjust required optional"
